pyproject.toml
[tool.pysen]
version = "0.10"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = true
mypy_preset = "strict"
line_length = 121
py_version = "py37"
[[tool.pysen.lint.mypy_targets]]
  paths = ["."]

[tool.poetry]
name = "utalib"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9.6"
requests = "^2.28.0"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pysen = {version = "^0.10.2", extras = ["lint"]}


[tool.black] # automatically generated by pysen
# pysen ignores and overwrites any modifications
line-length = 121
target-version = ["py37"]

[tool.isort] # automatically generated by pysen
# pysen ignores and overwrites any modifications
default_section = "THIRDPARTY"
ensure_newline_before_comments = true
force_grid_wrap = 0
force_single_line = false
include_trailing_comma = true
line_length = 121
multi_line_output = 3
use_parentheses = true
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
